Updated from Inno Setup 5.5.4 to 5.5.9
Just removed some directives where the new defaults are the right thing to do,
and there is no practical difference. Output file size is the same. What could
be done is incorporating this change from 5.5.7, but it looks like it could
require scripting to discern between an upgrade and a new installation:

"The Compiler IDE's New Script Wizard now offers to create a shortcut to the
main executable in the common Start Menu Programs folder instead of creating a
new Start Menu folder especially for the application. This option is enabled by
default and is recommended by Microsoft unless you install a suite of
applications rather than a single application."