Updated to Inno Setup 5.6.1, overhauled installer script
The installer compiler used is still the non-Unicode version. The installer is
a little smaller than before, due to different compressions settings.

Removed the script that moved data files from the installation folder to user-
specific locations when updating to MB 0.7 from an older version. This was
related to changes between Windows XP and Vista, and should no longer be
necessary.

A folder is no longer created in the Start menu, just an icon. This is per
Microsoft recommendations. The folder is not removed when updating from an
older version, but this should not cause any issues.

The option to create a shortcut in the Quick Launch area of the taskbar is
removed. This, together with the upgrade script removal, means that no user-
specific folders are written to.